Create generate_website_content.py
Copied from the demo repo. It should be called by a Github action, looks at the file database_information.csv and produces some files. The files are database_webpages/..., database_index.json. In the future, it might produce data used for a visualisation of what is present in the website.